pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sanctuary"
version = "0.1.0"
description = "Multi-agent LLM market simulation — The Sanctuary"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.7",
    "eval_type_backport>=0.2",   # Pydantic v2 union syntax on Python <3.10
    "anthropic>=0.40",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "PyYAML>=6.0",
    "numpy>=1.26",
    "matplotlib>=3.9",
    "reportlab>=4.2",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "websockets>=12.0",
    "aiofiles>=23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "black>=24.0",
    "pypdf>=4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sanctuary*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]

[tool.black]
line-length = 100
target-version = ["py312"]
